Added a timeout to the API request
At GameChanger, we've observed it's possible for our email queue processors to all simultaneously hang. When we restart them they are able to resume processing fine.

We believe that the lack of a timeout on the HTTP request to the Sailthru API is causing this request to never return in the event of a temporary network partition, causing the hanging of our processes.

This adds a reasonable timeout of 10 seconds to ensure that processes are able to recover.